BGV Relin Bug Fixes (#43)
This PR implements the an experimental fix for a few issues in the RELIN kernel. Namely, the RELIN kernel generated code that was segmented due issues with composing DigitalDecompExtend and Keymul together. This fix uses internal temporary variables output_tmp , which will be refactored out in a following PR. Further, the input polynomials parts are hardcoded to 3 and output parts to 2 since the code assumes those sizes (also to be refactored to be more generic in the future).

Today the RELIN kernel only supports 16k polynomials due to a limitation of the NTT/INTT kernels. This limitation is being removed and the RELIN kernel will be updated accordingly.